+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+- add posTimerCallbackSet to create a callback timer instead of normal one.
+  Callback is done from timer interrupt context with scheduler lock,
+  which limits it's usefulness (causes easily jitter in interrupt response).
 
 ## [1.1.1]
 - bug fixes to tickless idle